NetworkInterfaceInfo.Speed int --> long

+++ src/libraries/Common/src/Interop/Unix/System.Native/Interop.EnumerateInterfaceAddresses.cs
@@ -33,7 +33,7 @@ internal static partial class Interop
         {
             public fixed byte Name[16];
             public int InterfaceIndex;
-            public int Speed;
+            public long Speed;
             public int Mtu;
             public ushort HardwareType;
             public byte OperationalState;
